+++ src/main/java/com/rcextract/minecord/event/server/ServerCreateEvent.java
@@ -1,8 +1,12 @@
 package com.rcextract.minecord.event.server;
 
+import java.util.Arrays;
+import java.util.HashSet;
+import java.util.Set;
+
 import org.bukkit.event.HandlerList;
 
-import com.rcextract.minecord.ChannelManager;
+import com.rcextract.minecord.Channel;
 import com.rcextract.minecord.RankManager;
 import com.rcextract.minecord.Server;
 import com.rcextract.minecord.event.MinecordEvent;
@@ -20,15 +24,17 @@ public class ServerCreateEvent extends MinecordEvent {
 	private String desc;
 	private Boolean approvement;
 	private Boolean invitation;
-	private ChannelManager channelManager;
+	private final Set<Channel> channels;
+	private Channel main;
 	private RankManager rankManager;
-	public ServerCreateEvent(String name, String desc, Boolean approvement, Boolean invitation, ChannelManager channelManager, RankManager rankManager) {
+	public ServerCreateEvent(String name, String desc, Boolean approvement, Boolean invitation, RankManager rankManager, Channel main, Channel ... channels) {
 		this.name = name;
 		this.desc = desc;
 		this.approvement = approvement;
 		this.invitation = invitation;
-		this.channelManager = channelManager;
 		this.rankManager = rankManager;
+		this.main = main;
+		this.channels = new HashSet<Channel>(Arrays.asList(channels));
 	}
 	
 	@Override
@@ -60,16 +66,22 @@ public class ServerCreateEvent extends MinecordEvent {
 	public void setInvitation(Boolean invitation) {
 		this.invitation = invitation;
 	}
-	public ChannelManager getChannelManager() {
-		return channelManager;
-	}
-	public void setChannelManager(ChannelManager channelManager) {
-		this.channelManager = channelManager;
-	}
 	public RankManager getRankManager() {
 		return rankManager;
 	}
 	public void setRankManager(RankManager rankManager) {
 		this.rankManager = rankManager;
+	}
+
+	public Channel getMain() {
+		return main;
+	}
+
+	public void setMain(Channel main) {
+		this.main = main;
+	}
+
+	public Set<Channel> getChannels() {
+		return channels;
 	}
 }