+++ src/main/java/com/rcextract/minecord/utils/Pair.java
@@ -6,15 +6,21 @@ public class Pair<K, V> implements Map.Entry<K, V> {
 
 	private K key;
 	private V value;
+
+	public Pair() {}
+	
+	public Pair(K key, V value) {
+		this.key = key;
+		this.value = value;
+	}
+	
 	@Override
 	public K getKey() {
-		// TODO Auto-generated method stub
 		return key;
 	}
 
 	@Override
 	public V getValue() {
-		// TODO Auto-generated method stub
 		return value;
 	}